RFC: cri: allow setting supplementary groups via container annotations
CRI container runtimes mount devices (set via kubernetes device plugins)
to containers by taking the host owner/group IDs to the corresponding
container device.

This triggers a problem when trying to run those containers with
non-zero (root uid/gid = 0) uid/gid set via runAsUser/runAsGroup:
the container user has no permission to use the device even when
its gid is permissive to non-root users because the container user
does not belong to that group.

It is possible to workaround the problem by manually adding the device
gid(s) to supplementalGroups. However, this is also problematic because
the device gid(s) may have different values depending on the workers'
distro/version in the cluster.

This patch suggests to allow kubernetes device plugins to pass information
about worker/node specific host device gid(s) that need to be added to
supplementary groups via container annotations.

The idea follows the same principles with Kubernetes fsGroups: the
the container uid is added to those gid(s) with the difference that
host device permissions are not modified but simply taken to containers
as is.

TODO1: should kubelet do this?
TODO2: add device gids only if the securitycontext specifies non-root user

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>