cri: Devices ownership from SecurityContext
CRI container runtimes mount devices (set via kubernetes device plugins)
to containers by taking the host owner/group IDs to the corresponding
container device.

This triggers a problem when trying to run those containers with
non-zero (root uid/gid = 0) uid/gid set via runAsUser/runAsGroup:
the container user has no permission to use the device even when
its gid is permissive to non-root users because the container user
does not belong to that group.

It is possible to workaround the problem by manually adding the device
gid(s) to supplementalGroups. However, this is also problematic because
the device gid(s) may have different values depending on the workers'
distro/version in the cluster.

This patch suggests to take RunAsUser/RunAsGroup set via SecurityContext
as the device UID/GID, respectively. The feature can be opted-out by
setting io.kubernetes.cri.DisableDeviceOwnershipFromSecurityContext
annotation to "true".

The annotation can be set by kubernetes device plugins or kubelet.
If several device resources are requested, it is enough if one device
plugin sets this annotation to opt-out for the configuration for
*all* the devices added.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>